🗺️ Atlas: [extract query tests to submodules]
