+++ relvar-core/src/query/tests/mod.rs
@@ -1,0 +1,2 @@
+pub mod basic;
+pub mod common;
